Refactoring of code to have a new helper class.
